% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_species_data.R
\name{get_species}
\alias{get_species}
\title{Get species occurrence data}
\usage{
get_species(species, ext, limit = 500)
}
\arguments{
\item{species}{character name of the species}

\item{ext}{a vector  xmin,xmax,ymin,ymax}

\item{limit}{number of points to return}
}
\value{
returns a dataframe
}
\description{
Get species occurrence data
}
\examples{
get_species("Chromolaena odorata",ext=c(87,94,20,27),limit=10)

}
