% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bio_worldclim.R
\name{get_worldclim}
\alias{get_worldclim}
\title{Download data from worldclim_v_2.1}
\usage{
get_worldclim(name, wd = "./", ext = c(87, 94, 20, 27))
}
\arguments{
\item{name}{name of the data; valid arguments are "tmin", "tmax","tavg",
"precipitation", "srad","wspeed","wvapor", "bio","elevation". Check
worldclim website for more information.}

\item{wd}{defaults to current working directory; do not forget trailing slash}

\item{ext}{xmin,xmax,ymin,ymax; default set at c(88,90,20,25)}
}
\value{
A SpatRaster object cropped by ext
}
\description{
Download data from worldclim_v_2.1
}
\examples{

# Note:
# Currently on WorldClim v2.1 data are supported
# The raster package has a useful function getData() which performs much
# more than current functionality here. So, do check that one.

get_worldclim(name="elevation",wd="./",ext=c(87,94,20,27))

}
