% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_species_data.R
\name{multi_species}
\alias{multi_species}
\title{Combine multiple species occ data}
\usage{
multi_species(species, ...)
}
\arguments{
\item{species}{A list of species name}

\item{...}{Arguments from get_species function}
}
\value{
A list of data frames
}
\description{
Combine multiple species occ data
}
\examples{
# vector of species name
names <- c("Chromolaena odorata","Parthenium hysterophorus")
# extent bounds
ext=c(87,94,20,27)
multi_species(species=names,ext,limit=10)

}
